fix bugs

+++ examples/language_model/gpt-3/dygraph/run_pretrain.py
@@ -266,7 +266,7 @@ def do_train(args):
         # level O2 means converting the network to FP16
         if args.sharding_stage not in [2, 3]:
             scaler = fleet.distributed_scaler(scaler)
-        model = paddle.amp.decorate(models=model, level="O2", save_dtype="float32")
+        model = paddle.amp.decorate(models=model, level="O2")
 
     # wrap sharding stage2/3 and add collective group
     # TODO(Baibaifan): combine ShardingStage1/2/3 and fleet.distributed_model in feature
